requirements.txt
python-dotenv>=1.0.0
requests>=2.32.0
pandas>=2.2.0
numpy>=1.26.0
